scrapbook-classifier-service/requirements.txt
requests==2.25.1
pytest==6.2.2
Flask==1.1.2
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
flask_kafka==0.0.3
imageai==2.1.6
Pillow==8.1.2
python-dotenv==0.15.0
